requirements.txt
asgiref==3.7.2
DateTime==5.1
Django==3.2.7
django-restframework==0.0.1
django-utils==0.0.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.2
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
zope.interface==6.0
